src/requirements.txt
appdirs==1.4.4
backcall==0.2.0
brython==3.9.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
idna==2.10
iexfinance==0.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
pandas==1.2.3
pandas-datareader==0.9.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==2.0.1
